+++ br/pkg/restore/split/region.go
@@ -1,4 +1,4 @@
-// Copyright 2020 PingCAP, Inc. Licensed under Apache-2.0.
+// Copyright 2022 PingCAP, Inc. Licensed under Apache-2.0.
 
 package split
